aarch64: interactive graphical login desktop (#5, phase F — complete)
aarch64 now boots straight to the views graphical desktop: the compositor
forks vlogin, root/user authenticates against authd, and the session brings
up the taskbar (start menu + clock) over the desktop — the full #5 stack on
the LP64 port.

Cross-address-space buffer sharing (the client-window enabler):
 - sys/arch/aarch64/dev/display.c sys_shareregion (native 45): maps the
   compositor's window-buffer pages into the client's address space directly
   via pmap_extract (caller L1) -> pmap_map_user_page (client L1) at a VA
   bump-allocated from the client's mmap region.  No CR3 dance or PT self-map
   (the i386 vmm_share_region is 32-bit/x86-only); all RAM is identity-mapped
   so both page tables are walkable from the caller's context.
 - The share ABI's out_vaddr widened u_int32_t -> uintptr_t (sysproto.h, fb.c,
   window_manager.cc client_vaddr) so a 64-bit client VA survives; i386 is
   byte-identical (uintptr_t == u_int32_t there).  This also clears the
   int-to-pointer-truncation warnings the compositor had on aarch64.

Client apps + boot:
 - bin/Makefile + bin/views/Makefile build vlogin, taskbar and term on aarch64
   (all link as PIE via the phase-E C++ toolchain fix).
 - boot.c runs /bin/views as the disk-backed init (compositor -> vlogin) on the
   graphical path; the text /bin/login chain remains the console fallback.
 - mkimage-arm.sh stages the TrueType faces to /var/fonts (objGFX's scalable
   font backend — vlogin/taskbar load DejaVuSans*) and wallpapers to
   /var/background.

QEMU-verified (HVF, virtio-gpu/keyboard/mouse, QMP screendump + key injection):
boot -> "Do You Ubix?" graphical login renders (scalable-font UI over a
gradient); typing root <tab> user <enter> authenticates; vlogin spawns the
session and the taskbar (window 2, 1280x32 at the bottom) composites with a
working start menu + clock.  sys_shareregion hands each window a 64-bit shm
buffer (client VA 0x20002c000, confirmed matching compositor<->client).  i386
kernel + world build unchanged (ET_EXEC C++, empty PIE knobs, 32-bit share).

+++ sys/include/sys/sysproto.h
@@ -146,7 +146,7 @@ struct sys_shareregion_args
 	pid_t dst_pid;
 	void *vaddr;
 	u_int32_t size;
-	u_int32_t *out_vaddr;
+	uintptr_t *out_vaddr; /* register-width: 32-bit i386, 64-bit aarch64 VA */
 };
 int sys_shareregion(struct thread *, struct sys_shareregion_args *);
